refactor: DesignProvider 구조 변경
