+++ src/components/Input/Input.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import React, {forwardRef, useEffect, useImperativeHandle, useRef, useState} from 'react';
 
-import {useTheme} from '@theme/useTheme';
+import {useTheme} from '@theme/DesignProvider';
 
 import {Text} from '@components/Text';
 import {Icon} from '@components/Icon';
