+++ src/theme/DesignProvider.tsx
@@ -1,5 +1,5 @@
-import React, { ReactNode, useMemo} from 'react';
-import {Global, ThemeContext} from '@emotion/react';
+import React, { createContext, ReactNode, useContext, useMemo} from 'react';
+import {Global} from '@emotion/react';
 
 import {Theme} from '@theme/theme.type';
 import {GlobalStyle} from '@theme/GlobalStyle';
@@ -7,11 +7,17 @@ import {COLORS} from '@token/colors';
 import {TYPOGRAPHY} from '@token/typography';
 import {ZINDEX} from '@token/zIndex';
 
+interface ThemeContextProps {
+  theme: Theme;
+}
+
 const defaultTheme: Theme = {
   colors: COLORS,
   typography: TYPOGRAPHY,
   zIndex: ZINDEX,
 };
+
+export const ThemeContext = createContext<ThemeContextProps | undefined>(undefined);
 
 export const DesignProvider: React.FC<{children: ReactNode}> = ({children}) => {
   const theme = useMemo(() => defaultTheme, []);
@@ -22,4 +28,12 @@ export const DesignProvider: React.FC<{children: ReactNode}> = ({children}) => {
       {children}
     </ThemeContext.Provider>
   );
+};
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a DesignProvider');
+  }
+  return context;
 };
