--- src/theme/useTheme.ts
@@ -1,11 +1,0 @@
-import {useContext} from 'react';
-
-import {ThemeContext} from './ThemeContext';
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error('useTheme must be used within a DesignProvider');
-  }
-  return context;
-};
